Added more test coverage and separated unit tests from performance tests
