+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Scene/Messaging/MessagingBus.cs
@@ -216,7 +216,7 @@ namespace DCL
 
                             msgYieldInstruction = null;
                         }
-
+                        
                         break;
                     case QueuedSceneMessage.Type.LOAD_PARCEL:
                         handler.LoadParcelScenesExecute(m.message);
