+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Scene/PerformanceTests/PerformanceTests.asmdef
@@ -24,7 +24,8 @@
     "allowUnsafeCode": true,
     "overrideReferences": true,
     "precompiledReferences": [
-        "nunit.framework.dll"
+        "nunit.framework.dll",
+        "NSubstitute.dll"
     ],
     "autoReferenced": false,
     "defineConstraints": [
